requirements.txt
pandas
scikit-learn
scikit-surprise
flask
streamlit